created sun producing skill

+++ Plants vs Zombies 2 - Ap/src/models/npc/Sun.java
@@ -2,10 +2,10 @@ package models.npc;
 
 public class Sun {
     private int price;
-    private int remainingTime;
+    private float remainingTime;
     private int x;
     private int y;
-    public Sun(int price, int remainingTime, int x, int y) {}
+    public Sun(int price, float remainingTime, int x, int y) {}
 
     public void pickup(){}
     public Sun(int price, int remainingTime){}
